+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- added: New `enableRbf` field on `EdgeSpendInfo`
+- changed: Deprecated `rbfTxid` field on `EdgeSpendInfo`
+
 ## 2.6.1 (2024-05-23)
 
 - fixed: Add missing Swift import statement.